Update README.md A multidisciplinary Approach
#ioneurosciencehubfootnote.Intersection=)
Exploring the intersection of vitamin B12 and cognitive decline in older adults.This repository explores the complex interplay between the body and mind, highlighting the need for a holistic approach to health and wellness. We integrate knowledge from neuroscience, immunology, psychology, and osteopathy to foster research, discussion, and collaboration on the interconnectedness of human health. Disclaimer: This repository is for informational purposes only and does not provide medical advice.